Training can know be done against the new model
This made Mnesia raise it's head with overloaded dump_log, write_threshold because we now have much high parrallism on the training N classes means N concurrent with full term list. Seeing as it was swamped on the lineary this is not really a surprise. Solution could be modifying a RAM only and replicate back to a disk based table/node. This is though only a problem during training because classifying is read-only.