Got the math corrected for term counts in classes.
The implementation of vocabulary_size for class is extremely slow and quite the memory hog because it take all the term frequency for a  class into memory and runs it through. This number and the vocabulary size is actually constants as long as no documents are trained on the solution, which would be almost never compared to the number of classifications expected.